+++ legion/engine/llri-vk/llri.cpp
@@ -1,5 +1,8 @@
 #include <llri/llri.hpp>
 
+#define VOLK_IMPLEMENTATION
+#include <graphics/vulkan/volk.h>
+
 namespace legion::graphics::llri
 {
     //reserved for future use
